cloudformation/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
PyYAML>=6.0